pyproject.toml
[project]
name = "codewaifu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.0.4",
    "langchain-google-genai>=3.0.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
